% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrow.R
\name{row_concatenate_at}
\alias{row_concatenate_at}
\title{Concatenate selected columns rowwise and store the result in a
new column in the dataframe}
\usage{
row_concatenate_at(.tbl, .colname, .vars, sep = ", ", ...)
}
\arguments{
\item{.tbl}{A dataframe or tibble}

\item{.colname}{The name of the column to be added}

\item{.vars}{Columns to apply over}

\item{sep}{Separator for concatenation, defaults to ', '}

\item{...}{Additional arguments to str_c}
}
\description{
Concatenate selected columns rowwise and store the result in a
new column in the dataframe
}
