% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrow.R
\name{row_map_at}
\alias{row_map_at}
\title{Apply a function rowwise for selected variables in a table
and store the results in a new column}
\usage{
row_map_at(.tbl, .colname, .vars, .fun, ...)
}
\arguments{
\item{.tbl}{A dataframe or tibble}

\item{.colname}{The name of the column to be added}

\item{.vars}{Columns to apply over}

\item{.fun}{Function to apply}

\item{...}{Additional arguments to .fun}
}
\description{
Apply a function rowwise for selected variables in a table
and store the results in a new column
}
