% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrow.R
\name{row_mean_at}
\alias{row_mean_at}
\title{Average tidyselected columns rowwise and store the result as a
column in the dataframe}
\usage{
row_mean_at(.tbl, .colname, .vars, ...)
}
\arguments{
\item{.tbl}{A dataframe or tibble}

\item{.colname}{The name of the column to be added}

\item{.vars}{Columns to apply over}

\item{...}{Additional arguments to mean (e.g. na.rm = TRUE)}
}
\description{
Average tidyselected columns rowwise and store the result as a
column in the dataframe
}
