% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrow.R
\name{row_simple_scale}
\alias{row_simple_scale}
\title{Get a simple average scale score for a set of factor variables}
\usage{
row_simple_scale(.tbl, .colname, .vars, .rev, ...)
}
\arguments{
\item{...}{}
}
\description{
Get a simple average scale score for a set of factor variables
}
