% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrow.R
\name{row_sum_at}
\alias{row_sum_at}
\title{Sum tidyselected columns rowwise and store the result as a
column in the dataframe}
\usage{
row_sum_at(.tbl, .colname, .vars, ...)
}
\arguments{
\item{.tbl}{A dataframe or tibble}

\item{.colname}{The name of the column to be added}

\item{.vars}{Columns to apply over}

\item{...}{Additional arguments to sum (e.g. na.rm = TRUE)}
}
\description{
Sum tidyselected columns rowwise and store the result as a
column in the dataframe
}
